oracle/requirements.txt
torch
kafka-python
numpy
pyyaml
zstandard